Refactor of files structure in components directories
