+++ components/AccountLayout/index.ts
@@ -1,0 +1,2 @@
+export type { AccountLayoutProps } from "./AccountLayout";
+export { AccountLayout } from "./AccountLayout";
