+++ components/AddressBookCard/index.ts
@@ -1,0 +1,2 @@
+export type { AddressBookCardProps } from "./AddressBookCard";
+export { AddressBookCard } from "./AddressBookCard";
