+++ components/Button/index.ts
@@ -1,0 +1,2 @@
+export type { ButtonProps } from "./Button";
+export { Button } from "./Button";
