+++ components/CartSummary/index.tsx
@@ -1,0 +1,2 @@
+export type { CartSummaryProps } from "./CartSummary";
+export { CartSummary } from "./CartSummary";
