+++ components/CheckoutLineItem/index.ts
@@ -1,0 +1,1 @@
+export { CheckoutLineItem } from "./CheckoutLineItem";
