+++ components/HamburgerButton/index.ts
@@ -1,0 +1,2 @@
+export type { HamburgerButtonProps } from "./HamburgerButton";
+export { HamburgerButton } from "./HamburgerButton";
