+++ components/HomepageBlock/index.ts
@@ -1,0 +1,2 @@
+export type { HomepageBlockProps } from "./HomepageBlock";
+export { HomepageBlock } from "./HomepageBlock";
