+++ components/Layout/Layout.tsx
@@ -1,4 +1,4 @@
-import { Navbar } from ".";
+import { Navbar } from "../Navbar";
 
 export interface LayoutProps {
   children?: React.ReactNode;
