+++ components/Layout/index.ts
@@ -1,0 +1,2 @@
+export type { LayoutProps } from "./Layout";
+export { Layout } from "./Layout";
