+++ components/NavigationPanel/index.ts
@@ -1,0 +1,1 @@
+export { NavigationPanel } from "./NavigationPanel";
