+++ components/OrdersTable/index.ts
@@ -1,0 +1,2 @@
+export type { OrdersTableProps } from "./OrdersTable";
+export { OrdersTable } from "./OrdersTable";
