+++ components/PageHero/index.ts
@@ -1,0 +1,2 @@
+export type { PageHeroProps } from "./PageHero";
+export { PageHero } from "./PageHero";
