+++ components/Pagination/index.ts
@@ -1,0 +1,2 @@
+export type { PaginationProps } from "./Pagination";
+export { Pagination } from "./Pagination";
