+++ components/ProductCard/index.ts
@@ -1,0 +1,2 @@
+export type { ProductCardProps } from "./ProductCard";
+export { ProductCard } from "./ProductCard";
