+++ components/ProductCollection/index.ts
@@ -1,0 +1,2 @@
+export type { ProductCollectionProps } from "./ProductCollection";
+export { ProductCollection } from "./ProductCollection";
