+++ components/RichText/index.ts
@@ -1,0 +1,1 @@
+export { RichText } from "./RichText";
