+++ components/Spinner/index.ts
@@ -1,0 +1,1 @@
+export { Spinner } from "./Spinner";
